csvme hashify should strip leading & trailing whitespace

+++ lib/vmunger/array.rb
@@ -90,10 +90,11 @@ class Array
   # turns an array-of-arrays into an array-of-hashes
   # the headers are used as names for the fields
   def hashify(headers = shift)
+    hdrs = headers.map {|h| h && h.strip}
     select {|row| row.any?}.map do |row|
-      raise "Row count mismatch: #{row}" if row.count != headers.count
+      raise "Row count mismatch: #{row}" if row.count != hdrs.count
       hash = {}
-      row.zip(headers) {|v,k| hash[k] = v unless v.blank?}
+      row.zip(hdrs) {|v,k| hash[k] = v.strip unless v.blank?}
       # hash.delete_values(nil) # completely remove keys for nil values
       hash
     end
@@ -132,13 +133,27 @@ class Array
     end
   end
 
+  # marshal (ruby-specific binary format) the contents of this structure to a file
+  # fails if file exists
   def dumpme(filename)
     raise "#{filename} exists" if File.exists?(filename)
     File.open(filename, "w") {|f| f << Marshal.dump(self)}
   end
+  # same as #dumpme but overwrites existing file
   def dumpme!(filename)
-    File.unlink(filename)
+    File.unlink(filename) if File.exists?(filename)
     File.open(filename, "w") {|f| f << Marshal.dump(self)}
+  end
+
+  # attempt to dump out contents of this array-of-hashes as CSV to named file
+  # fields is list of attribute names to write out
+  # options headers is public names for the fields
+  def csvme(filename, fields, headers = fields)
+    CSV.open(filename, "wb") do |csv|
+      csv << headers unless headers.nil?
+      each {|hash| csv << fields.map {|f| hash[f]}}
+    end
+    true
   end
 
   # ARRAY OF HASHES
